+++ interactors/index.js
@@ -1,3 +1,4 @@
 export { default as HeadlineInteractor } from './headline';
 export { default as IconButtonInteractor } from './icon-button';
 export { default as DatepickerInteractor } from './datepicker';
+export { default as FormattedDateTimeInteractor } from './formatted-date-time';