+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for stripes-erm-testing
 
 ## 1.1.0 In progress
+  * ERM-2463 Added interactor for FormattedDateTime
 
 ## 1.0.0 2022-11-11
   * Added mock files